fix: close the four quality findings from the #1511-#1514 assessment
Four independent gaps surfaced by a /rhiza:quality run against ce3d8e1, none of
which any existing gate asked about.

**semgrep scanned nothing (#1511).** quality.mk still carried the pre-#1505
`[ -d $(SOURCE_FOLDER) ]` form, so on a repo with no src/ it warned and exited 0.
That is a silent pass in the mother repo, where rhiza_weekly.yml runs `make semgrep`
on a schedule. #1505 missed it because semgrep is owned by core rather than by
python.mk, and so was the guard: test_gate_scope.py's _SCOPED_GATES was a
hand-written list of four, and a gate absent from such a list cannot fail its own
regression test. semgrep now takes a SEMGREP_FOLDERS accumulator seeded from
SOURCE_FOLDER, bundles.mk contributes utils, and the list is no longer trusted --
test_every_declared_accumulator_is_guarded derives the expected set from the
`*_FOLDERS ?=` declarations in the bundles, so a sixth path-scoped gate cannot join
silently. `make semgrep` now scans 2 files with 5 rules.

**No executable documentation (#1512).** 100% interrogate coverage and zero doctest
examples: nothing verified that any docstring was still true. The shipped
.rhiza/tests/test_docstrings.py cannot reach utils/ for two independent reasons -- it
resolves SOURCE_FOLDER (no src/ here) and it walks packages by __init__.py (utils/
is flat) -- so rather than widen a test every consumer syncs, the mother repo brings
its own runner, as bundles.mk already does for the gate accumulators. 23 examples
across the four pure predicates, and test_module_carries_examples fails at zero,
since doctest.testmod reports success on a module with none.

**Test-layout parity undocumented (#1513).** The checker reported all 56 test files
as orphans of src/ modules that were never meant to exist. The by-behaviour layout
is deliberate but was recorded nowhere a tool could read, so the exemption had to be
re-derived by hand on every run. Now a [tool.check_test_layout] opt-out states it.

**Three modules below MI A (#1514).** test_bundle_content_validity.py (864 lines,
MI C 7.99), test_bundle_sync.py (838, B) and test_bundle_combinations.py (631, B)
split along the families they already had -- bodies moved verbatim. Every module in
the repo now ranks A. Two consequences the suite itself caught: the new generic
sweepers had to join _GENERIC_TESTS or a bundle merely swept by them would look
behaviourally covered, and the new conftest had to state its security posture. The
coverage meta-gate is verifiably no weaker -- no bundle is covered only by a file
this split created.

Gates: fmt, test (1517 passed, 45 skipped), typecheck, docs-coverage (100%), deps,
security, rhiza-test, semgrep, sync-self-check all green.

Closes #1511
Closes #1512
Closes #1513
Closes #1514
