+++ tests/bundles/test_bundle_combinations.py
@@ -1,326 +1,25 @@
-"""Multi-bundle sync combination tests.
+"""Byte-identity sync tests for bundles adopted together.
 
-Each test class simulates a distinct user intent — a profile or a hand-picked set
-of bundles — and asserts that the resulting project directory contains exactly the
-files the user would expect.
+What remains after #1514 split the module: the per-bundle and bundle-plus-overlay sync
+checks, each small and each asserting the same shape — sync this combination, and the
+files a consumer receives are byte-identical to the bundle sources.
+
+The CI-parity suites for the two hosted profiles moved to ``test_profile_ci_parity.py``
+and the dependency-group assertions to ``test_dependency_groups.py``.
+
+``TestVscodeBundleSync`` stays here deliberately: ``_EXEMPT['vscode']`` in
+``test_bundle_test_coverage.py`` names this module as where that bundle is covered.
 """
 
 from __future__ import annotations
 
 import json
-import re
-import tomllib
 from pathlib import Path
 
 import pytest
 import yaml
-from packaging.requirements import Requirement
 
 from tests.util import sync_bundles
-
-# Keyed by CI *job* name, valued by the make target it runs. Two rows where those
-# deliberately disagree, both for the same reason — the job name is a contract with
-# branch protection, the target name is not:
-#
-#   deptry     → `make deps`  the gate was renamed in #1474 (it was the one target whose
-#                             name differed by language); the job keeps its name because
-#                             "Check dependencies with deptry" is a required status check
-#                             in .github/rulesets/main-branch-protection.json
-#   pre-commit → `make fmt`   the runner became prek, the job name did not change
-PARITY_JOB_COMMANDS = {
-    "test": "make test",
-    "docs-coverage": "make docs-coverage",
-    "typecheck": "make typecheck",
-    "deptry": "make deps",
-    "pre-commit": "make fmt",
-    "security": "make security",
-    "license": "make license",
-}
-
-
-def _load_yaml(path: Path) -> dict:
-    """Load a YAML document from disk."""
-    with path.open(encoding="utf-8") as fh:
-        return yaml.safe_load(fh)
-
-
-def _normalise_job_name(name: str) -> str:
-    """Collapse forge-specific job namespaces to a shared parity key."""
-    return name.split(":", 1)[-1]
-
-
-def _supported_python_versions(root: Path) -> list[str]:
-    """Return the supported Python versions declared in pyproject classifiers."""
-    pyproject = (root / "pyproject.toml").read_text(encoding="utf-8")
-    return re.findall(r"Programming Language :: Python :: (3\.\d+)", pyproject)
-
-
-def _gitlab_jobs_from_includes(project: Path) -> dict[str, dict]:
-    """Load all included GitLab workflow jobs from the synced project."""
-    pipeline = _load_yaml(project / ".gitlab-ci.yml")
-    jobs: dict[str, dict] = {}
-
-    for include in pipeline.get("include", []):
-        local_path = include.get("local") if isinstance(include, dict) else None
-        if not local_path:
-            continue
-        # Honour GitLab `exists:` include rules: an include guarded by an
-        # `exists` clause is a no-op when the referenced file is absent (used by
-        # opt-in overlay bundles such as gitlab-quality-review). Skip it here so
-        # the profile-sync assertions match what GitLab would actually run.
-        skip = False
-        for rule in include.get("rules", []) if isinstance(include, dict) else []:
-            exists = rule.get("exists") if isinstance(rule, dict) else None
-            if exists and not any((project / candidate).exists() for candidate in exists):
-                skip = True
-                break
-        if skip:
-            continue
-        workflow = _load_yaml(project / local_path)
-        for name, job in workflow.items():
-            if isinstance(job, dict) and not name.startswith("."):
-                jobs[name] = job
-
-    return jobs
-
-
-def _job_commands(job: dict, key: str) -> list[str]:
-    """Return the shell commands for a GitHub or GitLab job block."""
-    commands = job.get(key, []) or []
-    if isinstance(commands, str):
-        return [commands]
-    return [str(command) for command in commands]
-
-
-def _contains_make_command(commands: str, target: str) -> bool:
-    """Return True when commands invoke the given Make target with or without -f."""
-    return any(candidate in commands for candidate in (f"make {target}", f"make -f .rhiza/rhiza.mk {target}"))
-
-
-def _group_has_dependency(group: list[str], package_name: str) -> bool:
-    """Return True if a dependency group contains a given package name (exact match)."""
-    return any(Requirement(dep).name.lower() == package_name.lower() for dep in group)
-
-
-# ---------------------------------------------------------------------------
-# Tests
-# ---------------------------------------------------------------------------
-
-
-def test_pyproject_declares_uv_dependency_groups(root: Path) -> None:
-    """pyproject.toml should expose test/docs uv groups for focused installs."""
-    with (root / "pyproject.toml").open("rb") as fh:
-        pyproject = tomllib.load(fh)
-
-    groups = pyproject.get("dependency-groups", {})
-    # 'lint' is deliberately not in this set (#1484) — prek/uvx provision every linter,
-    # so the group had nothing to hold and existed only as `lint = []`.
-    assert {"test", "docs"} <= set(groups)
-
-    # The test group declares exactly the third-party libraries `tests/` imports.
-    # python-dotenv used to sit here too, and was the clearest case of the drift this
-    # test now guards (#1484): nothing under tests/ imports it, and its one consumer —
-    # the shipped .rhiza/tests/test_docstrings.py — is run by `rhiza-test`, which
-    # injects `--with python-dotenv` itself.
-    assert _group_has_dependency(groups["test"], "pytest")
-    assert _group_has_dependency(groups["test"], "pyyaml")
-    assert _group_has_dependency(groups["test"], "defusedxml")
-    assert _group_has_dependency(groups["test"], "packaging")
-
-    # The docs group declares the notebook imports so deptry can resolve them.
-    assert _group_has_dependency(groups["docs"], "marimo")
-    assert _group_has_dependency(groups["docs"], "numpy")
-    assert _group_has_dependency(groups["docs"], "pandas")
-    assert _group_has_dependency(groups["docs"], "plotly")
-
-
-# Tools every make target provisions for itself via `uv run --with` / `uvx` (see
-# #1380). Declaring one in a dependency group is not a second opinion but a second
-# *source of truth*: the group's pin does not reach the `--with` resolution, so the
-# two silently disagree while the group still costs every `uv sync` a download. The
-# `typecheck` group held ty and mypy that way until #1484 — plus typer, whose
-# consumer had already been deleted in #1378.
-#
-# pytest's plugins are listed here but *pytest is deliberately not*, and the line is not
-# arbitrary: 48 of the files under tests/ contain `import pytest`, while none imports
-# pytest_cov, pytest_xdist, pytest_html, pytest_timeout or pytest_mock — those load
-# through entry points and no test ever names them. A library the suite imports has to
-# be in the project environment (a bare `uv run pytest tests/...` finds nothing
-# otherwise, and the shipped test_pyproject.py requires it); runner configuration does
-# not. Adding "pytest" below would fail the suite it is meant to protect.
-_PROVISIONED_ON_THE_FLY = frozenset(
-    {
-        "ty",
-        "mypy",
-        "typer",
-        "ruff",
-        "interrogate",
-        "pre-commit",
-        "prek",
-        "deptry",
-        "bandit",
-        "mutmut",
-        "hypothesis",
-        "pip-licenses",
-        "pytest-cov",
-        "pytest-xdist",
-        "pytest-html",
-        "pytest-timeout",
-        "pytest-mock",
-        "pytest-benchmark",
-    }
-)
-
-
-def test_no_dependency_group_declares_an_on_the_fly_tool(root: Path) -> None:
-    """No [dependency-groups] entry may duplicate a tool its make target already injects."""
-    with (root / "pyproject.toml").open("rb") as fh:
-        pyproject = tomllib.load(fh)
-
-    offenders = {
-        f"{group}:{Requirement(dep).name}"
-        for group, deps in pyproject.get("dependency-groups", {}).items()
-        for dep in deps
-        if Requirement(dep).name.lower() in _PROVISIONED_ON_THE_FLY
-    }
-    assert not offenders, (
-        "these dependency-group entries duplicate a tool provisioned by `uv run --with` "
-        f"or `uvx`; remove them and let the make target own the version: {sorted(offenders)}"
-    )
-
-
-def test_core_bundle_pyproject_declares_uv_dependency_groups(test_data_dir: Path) -> None:
-    """The starter pyproject fixture must expose the test/docs groups.
-
-    The path is tests/resources/pyproject.toml, not a bundle file: no bundle ships a
-    pyproject.toml at all, so this is the example skeleton rather than synced content.
-    'lint' left the expected set in #1484 along with the group itself.
-    """
-    pyproject_path = test_data_dir / "pyproject.toml"
-    assert pyproject_path.is_file(), f"{test_data_dir / 'pyproject.toml'} not found"
-
-    with pyproject_path.open("rb") as fh:
-        pyproject = tomllib.load(fh)
-
-    groups = pyproject.get("dependency-groups", {})
-    assert {"test", "docs"} <= set(groups), f"{pyproject_path} is missing dependency groups; found: {set(groups)}"
-
-
-class TestGitlabProjectProfileSync:
-    """The 'gitlab-project' profile produces CI pipelines but zero .github/ files."""
-
-    # Transitive closure of gitlab-project: gitlab-book, gitlab-marimo, gitlab-tests
-    # which require book, marimo, tests, gitlab, core
-    GITLAB_PROJECT_BUNDLES = [
-        "core",
-        "book",
-        "marimo",
-        "tests",
-        "gitlab",
-        "gitlab-book",
-        "gitlab-marimo",
-        "gitlab-tests",
-    ]
-
-    @pytest.fixture(autouse=True)
-    def synced(self, tmp_path: Path, root: Path) -> None:
-        """Sync the gitlab-project profile bundle closure into a fresh directory."""
-        sync_bundles(root, self.GITLAB_PROJECT_BUNDLES, tmp_path)
-        self.project = tmp_path
-
-    def test_gitlab_ci_file_exists(self) -> None:
-        """GitLab CI pipeline file must be present."""
-        assert (self.project / ".gitlab-ci.yml").is_file(), ".gitlab-ci.yml not found"
-
-    def test_gitlab_workflows_directory_exists(self) -> None:
-        """GitLab workflow helpers live under .gitlab/workflows/."""
-        assert (self.project / ".gitlab" / "workflows").is_dir()
-
-    def test_no_github_workflows_injected(self) -> None:
-        """GitLab profile must not produce any .github/workflows/ files."""
-        workflows = self.project / ".github" / "workflows"
-        if workflows.exists():
-            files = list(workflows.glob("*.yml"))
-            assert not files, f"Unexpected GitHub workflows in GitLab project: {[f.name for f in files]}"
-
-    def test_gitlab_ci_is_valid_yaml(self) -> None:
-        """The injected .gitlab-ci.yml must parse as valid YAML."""
-        content = (self.project / ".gitlab-ci.yml").read_text(encoding="utf-8")
-        parsed = yaml.safe_load(content)
-        assert parsed is not None, ".gitlab-ci.yml is empty or invalid YAML"
-
-    def test_core_infrastructure_present(self) -> None:
-        """Core Makefile and rhiza.mk must be present alongside GitLab CI."""
-        assert (self.project / "Makefile").is_file()
-        assert (self.project / ".rhiza" / "rhiza.mk").is_file()
-
-    def test_test_mk_present(self) -> None:
-        """test.mk fragment must be synced (from the tests bundle)."""
-        assert (self.project / ".rhiza" / "make.d" / "test.mk").is_file()
-
-    def test_ci_workflows_expose_same_core_job_names(self, root: Path) -> None:
-        """GitHub and GitLab CI definitions must expose the same core parity jobs."""
-        github_jobs = _load_yaml(root / ".github" / "workflows" / "rhiza_ci.yml")["jobs"]
-        gitlab_jobs = _gitlab_jobs_from_includes(self.project)
-        expected = set(PARITY_JOB_COMMANDS)
-
-        github_job_names = {_normalise_job_name(name) for name in github_jobs}
-        gitlab_job_names = {_normalise_job_name(name) for name in gitlab_jobs}
-
-        assert github_job_names & expected == expected
-        assert gitlab_job_names & expected == expected
-
-    def test_ci_workflows_share_python_version_matrix(self, root: Path) -> None:
-        """GitHub and GitLab CI must reference the same supported Python versions."""
-        github_workflow = _load_yaml(root / ".github" / "workflows" / "rhiza_ci.yml")
-        github_generate_steps = github_workflow["jobs"]["generate-matrix"]["steps"]
-        github_matrix_source = "\n".join(
-            step["run"] for step in github_generate_steps if isinstance(step, dict) and "run" in step
-        )
-        versions_step = next(step for step in github_generate_steps if step.get("id") == "versions")
-        gitlab_test_job = _gitlab_jobs_from_includes(self.project)["ci:test"]
-        gitlab_matrix = gitlab_test_job["parallel"]["matrix"][0]["PYTHON_VERSION"]
-
-        assert "hynek/build-and-inspect-python-package" in versions_step["uses"]
-        assert "supported_python_classifiers_json_array" in github_matrix_source
-        assert github_workflow["jobs"]["test"]["strategy"]["matrix"]["python-version"] == (
-            "${{ fromJson(needs.generate-matrix.outputs.matrix) }}"
-        )
-        assert gitlab_matrix == _supported_python_versions(root)
-
-    @pytest.mark.parametrize(("job_name", "command"), PARITY_JOB_COMMANDS.items())
-    def test_ci_workflows_match_core_commands(self, root: Path, job_name: str, command: str) -> None:
-        """GitHub and GitLab CI must run equivalent commands for core parity jobs."""
-        github_jobs = _load_yaml(root / ".github" / "workflows" / "rhiza_ci.yml")["jobs"]
-        gitlab_jobs = _gitlab_jobs_from_includes(self.project)
-
-        github_job = next(job for name, job in github_jobs.items() if _normalise_job_name(name) == job_name)
-        gitlab_job = next(job for name, job in gitlab_jobs.items() if _normalise_job_name(name) == job_name)
-
-        github_commands = "\n".join(
-            _job_commands(step, "run")[0] for step in github_job.get("steps", []) if "run" in step
-        )
-        gitlab_commands = "\n".join(_job_commands(gitlab_job, "script"))
-
-        target = command.removeprefix("make ")
-        assert _contains_make_command(github_commands, target)
-        assert _contains_make_command(gitlab_commands, target)
-
-    def test_gitlab_ci_jobs_define_timeout_budgets(self) -> None:
-        """GitLab CI core jobs must define explicit timeout budgets."""
-        gitlab_jobs = _gitlab_jobs_from_includes(self.project)
-        expected = {
-            "ci:test": "20m",
-            "ci:docs-coverage": "10m",
-            "ci:typecheck": "5m",
-            "ci:deptry": "5m",
-            "ci:pre-commit": "5m",
-            "ci:security": "10m",
-            "ci:license": "10m",
-        }
-
-        for job_name, timeout in expected.items():
-            assert gitlab_jobs[job_name]["timeout"] == timeout
 
 
 class TestDockerBundleSync:
@@ -544,88 +243,3 @@ class TestBenchmarksBundleSync:
         benchmarks_dir = self.project / "tests" / "benchmarks"
         has_setup = (benchmarks_dir / "conftest.py").is_file() or (benchmarks_dir / "__init__.py").is_file()
         assert has_setup, "tests/benchmarks/ has no conftest.py or __init__.py"
-
-
-class TestGithubProjectProfileSync:
-    """The 'github-project' profile produces a complete, CI-wired project skeleton."""
-
-    # Full transitive closure of github-project
-    GITHUB_PROJECT_BUNDLES = [
-        "core",
-        "book",
-        "marimo",
-        "tests",
-        "github",
-        "github-book",
-        "github-marimo",
-        "github-tests",
-    ]
-
-    @pytest.fixture(autouse=True)
-    def synced(self, tmp_path: Path, root: Path) -> None:
-        """Sync the full github-project bundle closure."""
-        sync_bundles(root, self.GITHUB_PROJECT_BUNDLES, tmp_path)
-        self.project = tmp_path
-
-    def test_ci_workflow_present(self) -> None:
-        """rhiza_ci.yml CI workflow must be present."""
-        assert (self.project / ".github" / "workflows" / "rhiza_ci.yml").is_file()
-
-    def test_codeql_workflow_present(self) -> None:
-        """rhiza_codeql.yml security workflow must be present."""
-        assert (self.project / ".github" / "workflows" / "rhiza_codeql.yml").is_file()
-
-    def test_book_workflow_present(self) -> None:
-        """rhiza_book.yml documentation workflow must be present."""
-        assert (self.project / ".github" / "workflows" / "rhiza_book.yml").is_file()
-
-    def test_benchmark_workflow_present(self) -> None:
-        """rhiza_benchmark.yml benchmark workflow must be present."""
-        assert (self.project / ".github" / "workflows" / "rhiza_benchmark.yml").is_file()
-
-    def test_marimo_workflow_present(self) -> None:
-        """rhiza_marimo.yml notebook workflow must be present."""
-        assert (self.project / ".github" / "workflows" / "rhiza_marimo.yml").is_file()
-
-    def test_all_injected_workflows_are_valid_yaml(self) -> None:
-        """Every injected workflow must be valid YAML with required fields.
-
-        Note: pyyaml parses 'on:' as Python boolean True (YAML boolean literal).
-        We check for both True and the string 'on' as valid trigger keys.
-        """
-        workflows_dir = self.project / ".github" / "workflows"
-        errors: list[str] = []
-        for wf in workflows_dir.glob("*.yml"):
-            with wf.open(encoding="utf-8") as fh:
-                doc = yaml.safe_load(fh)
-            if not isinstance(doc, dict):
-                errors.append(f"  {wf.name}: not a YAML mapping")
-                continue
-            # pyyaml parses 'on:' as True; check both forms
-            has_on = "on" in doc or any(k is True for k in doc)
-            if not has_on:
-                errors.append(f"  {wf.name}: missing 'on' trigger")
-            if "jobs" not in doc:
-                errors.append(f"  {wf.name}: missing 'jobs'")
-        if errors:
-            pytest.fail("Workflow validation failures:\n" + "\n".join(errors))
-
-    def test_no_gitlab_files_injected(self) -> None:
-        """github-project profile must not produce any GitLab CI files."""
-        assert not (self.project / ".gitlab-ci.yml").is_file(), (
-            ".gitlab-ci.yml should not be present in a github-project sync"
-        )
-        gitlab_dir = self.project / ".gitlab"
-        if gitlab_dir.exists():
-            # .gitlab/ dir might exist if some other mechanism wrote it; workflows should not
-            gitlab_wf = list((gitlab_dir / "workflows").glob("*.yml")) if (gitlab_dir / "workflows").exists() else []
-            assert not gitlab_wf, f"Unexpected GitLab workflows found: {[f.name for f in gitlab_wf]}"
-
-    def test_dependabot_present(self) -> None:
-        """dependabot.yml must be present from the github bundle."""
-        assert (self.project / ".github" / "dependabot.yml").is_file()
-
-    def test_docs_skeleton_present(self) -> None:
-        """docs/ directory with mkdocs base config must be present."""
-        assert (self.project / "docs" / "index.md").is_file()
-        assert (self.project / "docs" / "mkdocs-base.yml").is_file()
