+++ tests/bundles/test_bundle_sync.py
@@ -1,26 +1,22 @@
-"""Integration tests for the bundle-centric directory layout.
+"""Byte-identity sync tests for the core bundle, the Python layer and the profiles.
 
-Simulates what rhiza-cli does: resolves a set of bundle names, copies the
-files (following symlinks to get real content) into a fresh directory, and
-asserts that the resulting project looks exactly as expected.
+Reduced to this scope in #1514: the module had reached 838 lines at maintainability index
+**B (9.49)**, carrying the sync assertions for all three language layers plus the
+cross-layer contract in one file.
 
-Each test starts from a green-field temporary directory, so there is no
-shared state between scenarios.
+The Rust and Go layers moved to ``test_rust_layer_sync.py`` and ``test_go_layer_sync.py``
+— each is self-contained, since a layer's sync assertions never reference another's — and
+the properties that hold *across* layers moved to ``test_layer_contract.py``.
 """
 
 from __future__ import annotations
 
-import re
 import tomllib
 
 import pytest
 import yaml
 
 from tests.util import sync_bundles
-
-# ---------------------------------------------------------------------------
-# Tests
-# ---------------------------------------------------------------------------
 
 
 class TestCoreBundleSync:
@@ -184,474 +180,6 @@ class TestPythonCoreBundleSync:
         )
 
 
-class TestRustCoreBundleSync:
-    """Syncing core + rust-core produces a working Rust project layer.
-
-    The mother repo is a Python project, so nothing here is dogfooded at the root —
-    these assertions are the only thing standing between a broken rust-core file and
-    a downstream Rust repo.
-    """
-
-    @pytest.fixture(autouse=True)
-    def synced(self, tmp_path, root):
-        """Sync the core and rust-core bundles into a fresh directory."""
-        sync_bundles(root, ["core", "rust-core"], tmp_path)
-        self.project = tmp_path
-
-    @pytest.mark.parametrize(
-        "name", ["rust-toolchain.toml", "rustfmt.toml", "clippy.toml", "deny.toml", ".pre-commit-config.yaml"]
-    )
-    def test_rust_toolchain_config_is_present(self, name):
-        """The cargo-side counterparts of ruff.toml/.bandit all arrive."""
-        assert (self.project / name).is_file(), f"Missing Rust toolchain file: {name}"
-
-    @pytest.mark.parametrize("name", ["ruff.toml", ".bandit", ".python-version", "pytest.ini"])
-    def test_python_tooling_is_absent(self, name):
-        """A Rust repo gets no Python config — the layers are alternatives, not additions."""
-        assert not (self.project / name).exists(), f"{name} belongs to python-core, not rust-core"
-
-    def test_the_semgrep_config_arrives_without_a_python_bundle(self):
-        """`make semgrep` must be runnable on a Rust project (#1475).
-
-        The config moved from the `tests` bundle to `core` because `quality.mk` — core's —
-        is what names it. Before that a Rust sync got the target and not the file.
-        """
-        assert (self.project / ".rhiza" / "semgrep.yml").is_file(), (
-            "core's semgrep target has no config on a Rust project"
-        )
-
-    def test_rust_mk_provides_the_language_contract(self):
-        """rust.mk owns the same target names python.mk does, plus the cargo-backed gates.
-
-        ``rhiza-test`` is deliberately absent: its recipe is language-neutral, so it moved
-        to core's quality.mk in #1471 rather than existing identically in all three layers.
-        """
-        rust_mk = (self.project / ".rhiza" / "make.d" / "rust.mk").read_text(encoding="utf-8")
-        for target in ("install:", "all:", "test::", "coverage:", "typecheck:", "docs-coverage:"):
-            assert target in rust_mk, f"rust.mk is missing {target}"
-        assert "rhiza-test:" not in rust_mk, (
-            "rust.mk redefines rhiza-test, which core now owns — two recipes for one "
-            "target name is exactly what #1471 removed"
-        )
-
-    def test_only_one_language_fragment_is_synced(self):
-        """python.mk and rust.mk both define `install` — receiving both would be a clash."""
-        fragments = sorted(p.name for p in (self.project / ".rhiza" / "make.d").iterdir())
-        assert "rust.mk" in fragments
-        assert "python.mk" not in fragments
-
-    def test_no_symlinks_in_synced_project(self):
-        """Synced files are real files, as a downstream project would receive them."""
-        for path in self.project.rglob("*"):
-            if path.is_file():
-                assert not path.is_symlink(), f"Unexpected symlink in synced project: {path.relative_to(self.project)}"
-
-
-class TestNonPythonLayersShipADiscoverableBumpversionConfig:
-    """Rust and Go must land their version config where bump-my-version looks (#1453).
-
-    A Rust or Go repo has none of the four filenames bump-my-version searches, so
-    unlike Python it cannot be told to put the block in a file it already owns — the
-    layer has to ship one. That makes the placement load-bearing: the same config at
-    ``.rhiza/.cfg.toml`` (where these layers used to put it) is never read, and the
-    tool responds to finding no config by falling back to ``git describe`` instead of
-    failing. A release then gets cut against whatever the last reachable tag says.
-    """
-
-    LAYERS = ("rust-core", "go-core")
-
-    @pytest.fixture(params=LAYERS)
-    def synced(self, request, tmp_path, root):
-        """Sync one non-Python language layer and return (name, project dir)."""
-        sync_bundles(root, ["core", request.param], tmp_path)
-        return request.param, tmp_path
-
-    def test_the_config_is_at_a_discovered_path(self, synced):
-        """`.bumpversion.toml` is searched first; `.rhiza/.cfg.toml` is never searched."""
-        layer, project = synced
-        assert (project / ".bumpversion.toml").is_file(), f"{layer} ships no discoverable bumpversion config"
-        assert not (project / ".rhiza" / ".cfg.toml").exists(), (
-            f"{layer} still ships the inert .rhiza/.cfg.toml copy — bump-my-version never reads it"
-        )
-
-    def test_the_config_carries_no_current_version(self, synced):
-        """A synced file must not own a value only the consuming repo can maintain.
-
-        ``current_version`` would be reset to the template's number by every
-        ``/rhiza:update``. Leaving it out makes bump-my-version read the version from
-        the newest tag matching ``tag_name`` instead, which is what these layers want:
-        a Go module's version *is* its tag, and a Rust crate's Cargo.toml is expected
-        to agree with it.
-        """
-        layer, project = synced
-        with (project / ".bumpversion.toml").open("rb") as fh:
-            cfg = tomllib.load(fh)["tool"]["bumpversion"]
-        assert "current_version" not in cfg, f"{layer}: a synced config cannot own the repo's version"
-        assert cfg["tag_name"] == "v{new_version}", (
-            f"{layer}: tag_name is also the pattern the current version is read from, so it must "
-            f"match the tags the release flow creates"
-        )
-
-    @pytest.mark.parametrize("key", ["commit", "tag"])
-    def test_the_release_flow_owns_the_commit_and_the_tag(self, synced, key):
-        """`/rhiza:release` folds the changelog into the bump commit and tags it itself."""
-        layer, project = synced
-        with (project / ".bumpversion.toml").open("rb") as fh:
-            cfg = tomllib.load(fh)["tool"]["bumpversion"]
-        assert cfg[key] is False, (
-            f"{layer}: {key} = true makes a bare `bump-my-version bump` duplicate what the release flow already does"
-        )
-
-
-class TestRustBumpversionConfig:
-    """The Rust `.bumpversion.toml` must rewrite the crate version and nothing else.
-
-    This is the subtlest file rust-core ships. bump-my-version applies ``search`` to
-    *every* occurrence in a file, so the config anchors its pattern to the
-    ``[package]`` table — otherwise a dependency that happens to be pinned at the
-    same version as the crate gets silently rewritten too, and the release lands a
-    broken dependency spec. The comments in the file say so; this test proves it, by
-    running the config's own regex the way bump-my-version does.
-    """
-
-    CURRENT = "1.2.3"
-    NEW = "1.3.0"
-
-    CARGO_TOML = """\
-[package]
-name = "demo"
-version = "1.2.3"
-edition = "2024"
-
-[dependencies]
-serde = "1.2.3"
-
-[dependencies.tracing]
-version = "1.2.3"
-features = ["std"]
-"""
-
-    @pytest.fixture(autouse=True)
-    def config(self, tmp_path, root):
-        """Sync rust-core and load its bump-my-version configuration."""
-        sync_bundles(root, ["core", "rust-core"], tmp_path)
-        with (tmp_path / ".bumpversion.toml").open("rb") as fh:
-            self.cfg = tomllib.load(fh)["tool"]["bumpversion"]
-        self.project = tmp_path
-
-    def _bump(self) -> str:
-        """Apply the configured search/replace to CARGO_TOML exactly as bump-my-version would."""
-        spec = self.cfg["files"][0]
-        assert spec["filename"] == "Cargo.toml"
-        assert spec.get("regex") is True, "the [package] anchor is a regex; without regex=true it is a literal"
-        search = spec["search"].format(current_version=self.CURRENT)
-        replace = spec["replace"].format(new_version=self.NEW)
-        return re.sub(search, replace, self.CARGO_TOML)
-
-    def test_the_crate_version_is_bumped(self):
-        """The [package] version — the one the release is about — must change."""
-        result = self._bump()
-        assert f'[package]\nname = "demo"\nversion = "{self.NEW}"' in result
-
-    def test_a_dependency_pinned_at_the_same_version_is_untouched(self):
-        """The anchor exists for exactly this case; an unanchored search would rewrite it."""
-        result = self._bump()
-        assert 'serde = "1.2.3"' in result, "a shorthand dependency pin was rewritten"
-        assert f'[dependencies.tracing]\nversion = "{self.CURRENT}"' in result, (
-            "the version key of a [dependencies.*] table was rewritten — the [package] anchor is not holding"
-        )
-
-    def test_exactly_one_version_line_changes(self):
-        """Belt and braces: one occurrence of the new version, wherever the anchor matched."""
-        assert self._bump().count(f'"{self.NEW}"') == 1
-
-    def test_cargo_lock_is_refreshed_by_hook_not_by_search_replace(self):
-        """Cargo.lock lists every dependency's version, so it must never be a `files` entry."""
-        filenames = [f["filename"] for f in self.cfg["files"]]
-        assert "Cargo.lock" not in filenames, (
-            "Cargo.lock records dependency versions too; a search/replace entry would rewrite any "
-            "dependency pinned at the crate's version. Refresh it with `cargo update` instead."
-        )
-        hooks = " ".join(self.cfg.get("pre_commit_hooks", []))
-        assert "cargo update" in hooks, "nothing refreshes Cargo.lock, so the bumped tree dirties itself"
-        assert "git add Cargo.lock" in hooks, "the refreshed lockfile is never staged into the bump commit"
-
-    def test_prerelease_parts_are_semver_spellings(self):
-        """Cargo rejects PEP 440 short forms like `1.2.3-a.1`, so the values are spelled out."""
-        values = self.cfg["parts"]["release"]["values"]
-        assert "alpha" in values
-        assert "a" not in values
-
-
-class TestProfileRustLocalSync:
-    """Syncing the 'rust-local' profile yields a Rust project with no hosted CI."""
-
-    RUST_LOCAL_BUNDLES = ["core", "rust-core", "book"]  # transitive closure of the 'rust-local' profile
-
-    @pytest.fixture(autouse=True)
-    def synced(self, tmp_path, root):
-        """Sync all bundles in the rust-local profile transitive closure."""
-        sync_bundles(root, self.RUST_LOCAL_BUNDLES, tmp_path)
-        self.project = tmp_path
-
-    def test_the_language_contract_is_available(self):
-        """`install` and `all` exist, which is what makes the profile usable at all."""
-        rust_mk = (self.project / ".rhiza" / "make.d" / "rust.mk").read_text(encoding="utf-8")
-        assert "install:" in rust_mk
-        assert "all:" in rust_mk
-
-    def test_docs_skeleton_exists(self):
-        """The book bundle is language-neutral, so a Rust project gets docs too."""
-        assert (self.project / "docs" / "index.md").is_file()
-        assert (self.project / "docs" / "mkdocs-base.yml").is_file()
-
-    def test_book_badge_step_and_rust_coverage_agree_on_the_report_path(self):
-        """book.mk badges `_tests/coverage.xml`; the Rust `coverage` target must write it."""
-        book_mk = (self.project / ".rhiza" / "make.d" / "book.mk").read_text(encoding="utf-8")
-        rust_mk = (self.project / ".rhiza" / "make.d" / "rust.mk").read_text(encoding="utf-8")
-        assert "_tests/coverage.xml" in book_mk
-        assert "_tests/coverage.xml" in rust_mk
-
-    def test_no_github_workflows_injected(self):
-        """rust-local is local-first: the Rust CI workflows do not exist yet."""
-        workflows_dir = self.project / ".github" / "workflows"
-        if workflows_dir.exists():
-            workflow_files = list(workflows_dir.glob("*.yml"))
-            assert not workflow_files, (
-                f"rust-local should not inject workflow files, found: {[f.name for f in workflow_files]}"
-            )
-
-    def test_no_gitlab_ci_injected(self):
-        """Nor the GitLab half of hosted CI."""
-        assert not (self.project / ".gitlab-ci.yml").exists()
-
-    def test_exactly_one_pre_commit_config(self):
-        """Two layers would fight over this path; the profile selects exactly one."""
-        configs = list(self.project.rglob(".pre-commit-config.yaml"))
-        assert len(configs) == 1, f"expected one .pre-commit-config.yaml, found {configs}"
-        with configs[0].open(encoding="utf-8") as fh:
-            hook_ids = {hook["id"] for repo in yaml.safe_load(fh)["repos"] for hook in repo["hooks"]}
-        assert "cargo-clippy" in hook_ids, "the synced pre-commit config is not the Rust one"
-        assert not hook_ids & {"ruff", "bandit", "interrogate", "uv-lock"}, (
-            "python-core's hooks leaked into a Rust profile"
-        )
-
-
-class TestGoCoreBundleSync:
-    """Syncing core + go-core produces a working Go project layer.
-
-    Like rust-core, nothing here is dogfooded at the root — and unlike rust-core,
-    there is no Go toolchain in this environment at all, so these file-level
-    assertions are the whole safety net.
-    """
-
-    @pytest.fixture(autouse=True)
-    def synced(self, tmp_path, root):
-        """Sync the core and go-core bundles into a fresh directory."""
-        sync_bundles(root, ["core", "go-core"], tmp_path)
-        self.project = tmp_path
-
-    @pytest.mark.parametrize("name", [".golangci.yml", "revive.toml", ".pre-commit-config.yaml"])
-    def test_go_toolchain_config_is_present(self, name):
-        """The Go-side counterparts of ruff.toml/clippy.toml all arrive."""
-        assert (self.project / name).is_file(), f"Missing Go toolchain file: {name}"
-
-    @pytest.mark.parametrize("name", ["ruff.toml", ".bandit", ".python-version", "clippy.toml", "deny.toml"])
-    def test_sibling_layer_config_is_absent(self, name):
-        """A Go repo gets neither the Python nor the Rust toolchain — layers are alternatives."""
-        assert not (self.project / name).exists(), f"{name} belongs to another language layer"
-
-    def test_the_semgrep_config_arrives_without_a_python_bundle(self):
-        """`make semgrep` must be runnable on a Go project (#1475).
-
-        The config moved from the `tests` bundle to `core` because `quality.mk` — core's —
-        is what names it. Before that a Go sync got the target and not the file.
-        """
-        assert (self.project / ".rhiza" / "semgrep.yml").is_file(), (
-            "core's semgrep target has no config on a Go project"
-        )
-
-    def test_go_mk_provides_the_language_contract(self):
-        """go.mk owns the same target names its siblings do, plus the go-backed gates.
-
-        ``rhiza-test`` is deliberately absent: its recipe is language-neutral, so it moved
-        to core's quality.mk in #1471 rather than existing identically in all three layers.
-        """
-        go_mk = (self.project / ".rhiza" / "make.d" / "go.mk").read_text(encoding="utf-8")
-        for target in ("install:", "all:", "test::", "coverage:", "typecheck:", "docs-coverage:"):
-            assert target in go_mk, f"go.mk is missing {target}"
-        assert "rhiza-test:" not in go_mk, (
-            "go.mk redefines rhiza-test, which core now owns — two recipes for one "
-            "target name is exactly what #1471 removed"
-        )
-
-    def test_only_one_language_fragment_is_synced(self):
-        """python.mk, rust.mk and go.mk all define `install` — receiving two would be a clash."""
-        fragments = sorted(p.name for p in (self.project / ".rhiza" / "make.d").iterdir())
-        assert "go.mk" in fragments
-        assert "python.mk" not in fragments
-        assert "rust.mk" not in fragments
-
-    def test_the_version_constant_ships_with_the_layer(self):
-        """Unlike Cargo.toml, this file does not exist in a Go project unless rhiza puts it there."""
-        version_go = self.project / "internal" / "version" / "version.go"
-        assert version_go.is_file(), (
-            "go-core must ship internal/version/version.go — a Go module has no manifest, so "
-            "without it the release flow has no version location to write to"
-        )
-        source = version_go.read_text(encoding="utf-8")
-        assert source.startswith("// Package version"), "the package doc comment the docs-coverage gate requires"
-        assert 'const Version = "0.0.0"' in source
-
-    def test_the_layer_ships_a_test_so_the_test_gate_is_not_vacuous(self):
-        """`go test ./...` exits 0 on a package with no test file, printing "[no test files]".
-
-        So a freshly synced Go project passed `make test` — and `make all` — while
-        running nothing at all. Rust never had this hole: `cargo init --lib` leaves an
-        `it_works` test behind, while `go mod init` writes no Go file whatsoever, so
-        the layer has to bring the first test itself.
-        """
-        version_test = self.project / "internal" / "version" / "version_test.go"
-        assert version_test.is_file(), (
-            "go-core must ship internal/version/version_test.go — without a single test "
-            "file a fresh Go repo's `make test` is green without testing anything"
-        )
-        source = version_test.read_text(encoding="utf-8")
-        assert "func Test" in source, "the file ships no test function, so it does not close the vacuum"
-
-        # Comments are stripped before the literal check: the file explains in prose why
-        # it does not compare against "0.0.0", and matching that would fail the test for
-        # documenting the very trap it avoids.
-        code = "\n".join(line for line in source.splitlines() if not line.lstrip().startswith("//"))
-        assert '"0.0.0"' not in code, (
-            "the test must not compare against the literal shipped version: bump-my-version "
-            "rewrites version.go and not this file, so a literal turns red on the first release"
-        )
-
-    def test_no_symlinks_in_synced_project(self):
-        """Synced files are real files, as a downstream project would receive them."""
-        for path in self.project.rglob("*"):
-            if path.is_file():
-                assert not path.is_symlink(), f"Unexpected symlink in synced project: {path.relative_to(self.project)}"
-
-
-class TestGoBumpversionConfig:
-    """The Go `.bumpversion.toml` must write the one version location a Go module has.
-
-    Go is the odd layer out: pyproject.toml and Cargo.toml carry a version, a Go
-    module does not — its version is the git tag. The layer ships a `Version`
-    constant to serve as that location, so this checks the release config and the
-    shipped file agree, by running the config's own search the way bump-my-version
-    does. A rename on either side breaks the release with `ignore_missing_files
-    = false`, and this is what catches it first.
-    """
-
-    CURRENT = "1.2.3"
-    NEW = "1.3.0"
-
-    @pytest.fixture(autouse=True)
-    def config(self, tmp_path, root):
-        """Sync go-core and load its bump-my-version configuration."""
-        sync_bundles(root, ["core", "go-core"], tmp_path)
-        with (tmp_path / ".bumpversion.toml").open("rb") as fh:
-            self.cfg = tomllib.load(fh)["tool"]["bumpversion"]
-        self.project = tmp_path
-
-    def test_the_configured_file_is_the_file_the_bundle_ships(self):
-        """`ignore_missing_files = false`, so a path typo fails the release, not a test."""
-        assert self.cfg["ignore_missing_files"] is False
-        filenames = [f["filename"] for f in self.cfg["files"]]
-        assert filenames == ["internal/version/version.go"]
-        assert (self.project / filenames[0]).is_file(), "the config points at a file the layer does not ship"
-
-    def test_the_search_matches_the_shipped_constant(self):
-        """The declaration in version.go and the pattern in .bumpversion.toml must stay in step."""
-        spec = self.cfg["files"][0]
-        source = (self.project / spec["filename"]).read_text(encoding="utf-8")
-        # Version numbers are literal here; substitute the shipped 0.0.0 to match.
-        assert spec["search"].format(current_version="0.0.0") in source, (
-            "the bump-my-version search does not match the constant in version.go — "
-            "the release would fail with 'did not find current version'"
-        )
-
-    def test_a_version_in_prose_is_not_rewritten(self):
-        """Anchoring on `const Version =` is what keeps a doc comment or example safe."""
-        spec = self.cfg["files"][0]
-        search = spec["search"].format(current_version=self.CURRENT)
-        replace = spec["replace"].format(new_version=self.NEW)
-        source = f'// Version was {self.CURRENT} in the last release.\nconst Version = "{self.CURRENT}"\n'
-        result = source.replace(search, replace)
-        assert f"// Version was {self.CURRENT} in the last release." in result
-        assert f'const Version = "{self.NEW}"' in result
-
-    def test_no_lockfile_hook_is_needed(self):
-        """go.sum records dependency checksums only, so there is nothing to refresh."""
-        assert not self.cfg.get("pre_commit_hooks"), (
-            "the Go layer needs no post-bump hook; go.sum never records the module's own version"
-        )
-
-    def test_the_tag_is_the_semver_go_expects(self):
-        """Go resolves module versions from `vX.Y.Z` tags, and rejects PEP 440 short forms."""
-        assert self.cfg["tag_name"] == "v{new_version}"
-        values = self.cfg["parts"]["release"]["values"]
-        assert "alpha" in values
-        assert "a" not in values
-
-
-class TestProfileGoLocalSync:
-    """Syncing the 'go-local' profile yields a Go project with no hosted CI."""
-
-    GO_LOCAL_BUNDLES = ["core", "go-core", "book"]  # transitive closure of the 'go-local' profile
-
-    @pytest.fixture(autouse=True)
-    def synced(self, tmp_path, root):
-        """Sync all bundles in the go-local profile transitive closure."""
-        sync_bundles(root, self.GO_LOCAL_BUNDLES, tmp_path)
-        self.project = tmp_path
-
-    def test_the_language_contract_is_available(self):
-        """`install` and `all` exist, which is what makes the profile usable at all."""
-        go_mk = (self.project / ".rhiza" / "make.d" / "go.mk").read_text(encoding="utf-8")
-        assert "install:" in go_mk
-        assert "all:" in go_mk
-
-    def test_docs_skeleton_exists(self):
-        """The book bundle is language-neutral, so a Go project gets docs too."""
-        assert (self.project / "docs" / "index.md").is_file()
-        assert (self.project / "docs" / "mkdocs-base.yml").is_file()
-
-    def test_book_badge_step_and_go_coverage_agree_on_the_report_path(self):
-        """book.mk badges `_tests/coverage.xml`; the Go `coverage` target must write it."""
-        book_mk = (self.project / ".rhiza" / "make.d" / "book.mk").read_text(encoding="utf-8")
-        go_mk = (self.project / ".rhiza" / "make.d" / "go.mk").read_text(encoding="utf-8")
-        assert "_tests/coverage.xml" in book_mk
-        assert "_tests/coverage.xml" in go_mk
-
-    def test_no_github_workflows_injected(self):
-        """go-local is local-first: the Go CI workflows do not exist yet."""
-        workflows_dir = self.project / ".github" / "workflows"
-        if workflows_dir.exists():
-            workflow_files = list(workflows_dir.glob("*.yml"))
-            assert not workflow_files, (
-                f"go-local should not inject workflow files, found: {[f.name for f in workflow_files]}"
-            )
-
-    def test_no_gitlab_ci_injected(self):
-        """Nor the GitLab half of hosted CI."""
-        assert not (self.project / ".gitlab-ci.yml").exists()
-
-    def test_exactly_one_pre_commit_config(self):
-        """Three layers now claim this path; the profile selects exactly one."""
-        configs = list(self.project.rglob(".pre-commit-config.yaml"))
-        assert len(configs) == 1, f"expected one .pre-commit-config.yaml, found {configs}"
-        with configs[0].open(encoding="utf-8") as fh:
-            hook_ids = {hook["id"] for repo in yaml.safe_load(fh)["repos"] for hook in repo["hooks"]}
-        assert "go-mod-tidy" in hook_ids, "the synced pre-commit config is not the Go one"
-        assert not hook_ids & {"ruff", "bandit", "interrogate", "uv-lock", "cargo-clippy"}, (
-            "a sibling layer's hooks leaked into a Go profile"
-        )
-
-
 class TestCoreAndTestsBundleSync:
     """Syncing the Python layer plus the tests bundle adds pytest infrastructure.
 
@@ -749,90 +277,3 @@ class TestProfileLocalSync:
             assert not workflow_files, (
                 f"Local profile should not inject workflow files, found: {[f.name for f in workflow_files]}"
             )
-
-
-@pytest.mark.parametrize("layer", ["python-core", "rust-core", "go-core"])
-class TestALayersAllIsSatisfiableOnItsOwn:
-    """`core` + one language layer must be enough to run that layer's `all`.
-
-    This is the invariant #1475 established. `python.mk`'s `all` named `test`,
-    `typecheck`, `security` and `docs-coverage` while the `tests` bundle defined them,
-    and nothing made `tests` arrive — the dependency runs the other way. So
-    `core + python-core` had an `all` that died on a missing rule, while the Rust and Go
-    layers were self-contained. No shipped profile reached it (all three Python profiles
-    select `tests`), which is exactly why it went unnoticed; a hand-written
-    `.rhiza/template.yml` of `[core, python-core]` did.
-
-    Parametrised over all three layers rather than written for Python, because the point
-    is the property, not the one instance of it that was broken.
-    """
-
-    def test_every_prerequisite_of_all_has_a_rule(self, tmp_path, root, layer):
-        """Each name in `all`'s prerequisite list is defined by some synced fragment."""
-        sync_bundles(root, ["core", layer], tmp_path)
-        fragments = list((tmp_path / ".rhiza").rglob("*.mk"))
-        assert fragments, f"{layer} synced no make fragments"
-
-        text = {f: f.read_text(encoding="utf-8") for f in fragments}
-
-        all_line = next(
-            (line for body in text.values() for line in body.splitlines() if line.startswith("all:")),
-            None,
-        )
-        assert all_line, f"{layer} defines no `all` target"
-        prerequisites = all_line.split("##")[0].split(":", 1)[1].split()
-
-        defined = {
-            line.split(":", 1)[0].rstrip(":")
-            for body in text.values()
-            for line in body.splitlines()
-            if re.match(r"^[a-z][a-z0-9-]*::? ", line) or re.match(r"^[a-z][a-z0-9-]*::?$", line)
-        }
-
-        missing = [name for name in prerequisites if name not in defined]
-        assert not missing, (
-            f"`make all` on core + {layer} names {missing}, which no synced fragment defines. "
-            f"Either the layer must define them or they belong in core — a gate that only "
-            f"resolves once another bundle happens to be selected is not part of the contract."
-        )
-
-
-class TestEveryLayerDefinesTheSameGateNames:
-    """The layers must agree on gate *names*, whatever engine backs each one.
-
-    The whole claim of the language-layer split is "same target names, different
-    recipes" — that is what lets book.mk, the CI workflows and the docs call a gate
-    without knowing the language. `deptry` was the one row that broke it (#1474):
-    python-core named the unused-dependency gate after the tool while rust-core and
-    go-core called it `deps`, so `make deps` failed on Python and `make deptry` on the
-    other two, and no language-neutral caller could invoke it at all.
-
-    CLAUDE.md documented the divergence, which is not the same as it being intended.
-    This test is why the next one fails instead of getting written down.
-    """
-
-    # Deliberately not `install`/`all` — `tests/api/test_language_layer.py` covers those
-    # behaviourally. These are the gates a caller reaches for by name.
-    GATES = ("test", "typecheck", "security", "docs-coverage", "license", "deps")
-
-    @pytest.mark.parametrize("gate", GATES)
-    def test_all_three_layers_define_the_gate(self, tmp_path, root, gate):
-        """Each gate name must be defined by every language layer."""
-        missing = []
-        for layer in ("python-core", "rust-core", "go-core"):
-            project = tmp_path / layer
-            sync_bundles(root, ["core", layer], project)
-            defined = {
-                line.split(":", 1)[0]
-                for fragment in (project / ".rhiza").rglob("*.mk")
-                for line in fragment.read_text(encoding="utf-8").splitlines()
-                if re.match(r"^[a-z][a-z0-9-]*::? ", line)
-            }
-            if gate not in defined:
-                missing.append(layer)
-
-        assert not missing, (
-            f"`{gate}` is not defined by {missing}. A gate the other layers provide under "
-            f"this name makes the contract language-specific: a caller has to know what the "
-            f"project is written in before it can run the gate."
-        )
